Changed ID type to BIGINT

+++ database/migrations/2025_06_17_162408_create_admins_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::disableForeignKeyConstraints();
 
         Schema::create('admins', function (Blueprint $table) {
-            $table->integer('id')->primary();
+            $table->bigInteger('id')->primary();
             $table->string('first_name');
             $table->string('last_name');
             $table->string('email')->unique();
